renaming authentication view by admin view + move addContract function into it
