+++ front/src/layout/header/MainHeader.vue
@@ -1,7 +1,7 @@
 <template>
     <b-navbar toggleable="lg" type="dark" variant="dark" fixed="top">
         <b-navbar-brand to="/">
-            Home
+            Admin
         </b-navbar-brand>
 
         <b-navbar-toggle target="nav-collapse"></b-navbar-toggle>
