--- front/src/pages/Authentication/Authentication.vue
@@ -1,12 +1,0 @@
-<template>
-    <h1>Authentication</h1>
-</template>
-<style scoped src="./Authentication.css">
-
-</style>
-<script lang="ts">
-import { Component, Prop, Vue } from 'vue-property-decorator';
-@Component
-export default class Authentication extends Vue {
-}
-</script>
